add: 教材削除確認ダイアログをコンポーネント化
