+++ api/app/models/user.rb
@@ -10,7 +10,7 @@ class User < ActiveRecord::Base
   has_many :likes, dependent: :destroy
 
   before_save { self.email = email.downcase }
-  validates :name,  presence: true, length: { maximum: 50 }
+  validates :name,  presence: true, length: { maximum: 50 }, uniqueness: true
   VALID_EMAIL_REGEX = /\A[\w+\-.]+@[a-z\d\-.]+\.[a-z]+\z/i
   validates :email, presence: true, length: { maximum: 255 },
                     format: { with: VALID_EMAIL_REGEX },
